No error
Changed the "check" command to output warnings to the log file rather than to throw an exception when problems are detected. If an exception is thrown, all of the entries in the properties file are not checked. Also, removed the CV repetition since the errors will not change with different iterations.